[Feat] 설정탭에서 dateTimeSetting 구현

+++ FreshNote/Domain/UseCases/UpdatePushNotificationUseCase.swift
@@ -18,19 +18,34 @@ protocol UpdatePushNotificationUseCase {
   /// 전 제품 알림 업데이트(d-day 변경되면 반드시 호출되는 메소드)
   /// dateTime이 변경될 때 호출되는 메소드
   /// 이 메소드 내에서 검증 실패 시, 알림을 삭제할 수도 있음
-  func updateNotifications(products: [Product]) -> AnyPublisher<Void, any Error>
+  func updateNotifications() -> AnyPublisher<Void, any Error>
 }
 
 final class DefaultUpdatePushNotificationUseCase: UpdatePushNotificationUseCase {
   private let savePushNotificationUseCase: any SavePushNotificationUseCase
   private let deletePushNotificationUseCase: any DeletePushNotificationUseCase
   
+  private let fetchProductUseCase: (any FetchProductUseCase)?
+  
+  /// 단일 로컬 알림 update
   init(
     savePushNotificationUseCase: any SavePushNotificationUseCase,
     deletePushNotificationUseCase: any DeletePushNotificationUseCase
   ) {
     self.savePushNotificationUseCase = savePushNotificationUseCase
     self.deletePushNotificationUseCase = deletePushNotificationUseCase
+    self.fetchProductUseCase = nil
+  }
+  
+  /// 모든 로컬 알림 update
+  init(
+    savePushNotificationUseCase: any SavePushNotificationUseCase,
+    deletePushNotificationUseCase: any DeletePushNotificationUseCase,
+    fetchProductUseCase: any FetchProductUseCase
+  ) {
+    self.savePushNotificationUseCase = savePushNotificationUseCase
+    self.deletePushNotificationUseCase = deletePushNotificationUseCase
+    self.fetchProductUseCase = fetchProductUseCase
   }
   
   func updateNotification(product: Product) -> AnyPublisher<Void, any Error> {
@@ -41,21 +56,31 @@ final class DefaultUpdatePushNotificationUseCase: UpdatePushNotificationUseCase 
       .saveNotification(product: product)
   }
   
-  func updateNotifications(products: [Product]) -> AnyPublisher<Void, any Error> {
-    let productIDs = products.map { $0.did }
-    self.deletePushNotificationUseCase
-      .deleteAllNotifications(productIDs: productIDs)
+  func updateNotifications() -> AnyPublisher<Void, any Error> {
+    guard let fetchProductUseCase = self.fetchProductUseCase else {
+      return Empty<Void, any Error>().eraseToAnyPublisher()
+    }
     
-    return Publishers
-      .Sequence(sequence: products)
-      .flatMap(maxPublishers: .max(3)) { [weak self] product -> AnyPublisher<Void, any Error> in
-        guard let self else { return Empty<Void, any Error>().eraseToAnyPublisher() }
+    return fetchProductUseCase
+      .fetchProducts()
+      .flatMap { [weak self] products -> AnyPublisher<Void, any Error> in
+        guard let self else { return Empty().eraseToAnyPublisher() }
+        let productIDs = products.map { $0.did }
+        self.deletePushNotificationUseCase
+          .deleteAllNotifications(productIDs: productIDs)
         
-        return self.savePushNotificationUseCase
-          .saveNotification(product: product)
+        return Publishers
+          .Sequence(sequence: products)
+          .flatMap(maxPublishers: .max(3)) { [weak self] product -> AnyPublisher<Void, any Error> in
+            guard let self else { return Empty<Void, any Error>().eraseToAnyPublisher() }
+            
+            return self.savePushNotificationUseCase
+              .saveNotification(product: product)
+          }
+          .collect()
+          .map { _ in }
+          .eraseToAnyPublisher()
       }
-      .collect()
-      .map { _ in }
       .eraseToAnyPublisher()
   }
 }
